requirements.txt
fastapi[all]==0.110.0
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.6.3
uvicorn==0.27.1
bcrypt==4.0.1
